Font styling for code blocks.
For code and other elements that use a fixed width font, add Menlo to
the top of the font stack, as almost all Macs have it installed.
Everything else will get Courier.

Reduce the font size to show more of the code, and to accommodate for the
larger Menlo font.