Indent lists the same as paragraphs.
`ol` and `ul` elements were indented the default `40px` in padding, and have
position set to `outside`, which doesn't include the markers as part of the box
model.

This results in the indent being about half what the paragraph indent is.

Set the indent specifically to the same number as paragraphs, and change the
position to be `inside`, which counts markers as part of the box model.