Improve paragraph indentation code.
The paragraph indent code was very specific and verbose, removing the indent
after headers.

Change the indent code to only apply to `p` tags that follow other `p` tags,
which is a nice, idiomatic rule that consistently applies we want. It also has
the added benefit of removing the indentation for paragraphs that follow other
block level elements, which have whitespace applied around them and signal a
break.