Add basic deploy on push script
Moving to a model where I deploy the site on a git push to the master
branch, for increased flexibility and to enable multiple codebases on
the same server.

Right now, the script responds correctly to a github push, has some
error handling, and verifies that the signature of the push is correct.

The next stage will be to actually build the site, and to decide which
site to build based on the github deploy payload.

Reject pushes that aren't to the master branch

Build sites dynamically

Which site this script will build is now determined in the
credentials.json file. The format is:

```
{
  "repository/full_name": {
    "secret": "my-super-secret",
    "path": "path/to/repo/on/server"
  }
}
```

Properly return on errors

HTTP error responses would not return, so the program would continue to
run, eventually hitting something it could not process and crashing.

This fix simply returns those errors to close the program loop,
preventing crashes.

Add proxy rule for deploy endpoint

The deploy endpoint is a node script that runs on localhost on the
server. This proxy allows it to be accessed from the open internet.