Switch to interal Chroma syntax highlighting
Instead of having to rely on the external python Pygments llibrary, the
new Hugo bundles the Chroma library, which is syntax highlighting in Go.

Use this new syntax highlighter, and switch the theme for ease of use.

This should fix my ongoing issues with node not being able to run syntax
highlighting when I deploy code automatically, since the syntax
highlighter is internal now.