Remove about me SVG from <p> tag
The about me graphic sits inside it's own little `<p>` tag, which in my
style gets a text indent. On small screens that pushes the graphic off
the screen.

Remove the space between the SVG and the previous paragraph in the
markdown file, which puts the SVG inside the previous `<p>` tag, solving
the problem, in this case at least.