Concat normalize as separate file
Separate out normalize.css into it's own file again, and concatenate it
with the theme.css on build.

This keeps normalize contained, in case I want to upgrade it in the
future, but still keeps the styles in one asset bundle for performance.

It also allows me to segment out some core styles into their own project
and concat that back into one theme.css.