Add static webmentions
Use static webmention data to markup and style each comment.

Put a comment and like counter at the bottom of articles in the form of
a partial, so it can be re-used across content types.

Move the webmention API call to the single page, so it can be passed
into both partials that need it's data, instead of making a separate API
call from within each partial.

Simplify webmention javascript. Even though this solution doesn't use
javascript, there's a case for using it to fetch more recent comments
then what's possible statically.