Game UX enhancements
Scroll to the beginning of the story when the player restarts. I think
this is less disorienting, and more what would be expected to happen.

Change to showing selected choices with a color in a class, rather than
an emoji. The emoji shows whether you made a good choice or a bad
choice, depending on the symbol, or the symbol is meaningless and
confusing. We just want to show the player which option they chose.