Fix game path bug
If the player gets an achievement for a path, that path's button then
becomes no longer opens up that path after a restart.

This is because the achievement code includes a return inside an if
condition, which breaks out of the surrounding function entirely before
the code that reveals the next path gets a chance to run.