Correct sync command
The problem with the strange permissions issues was that I didn't have
permissions correctly set on the server AND my rsync command was off.

I changed all permissions on directories to 750, and files to 640, which
seems sane. I also had to make sure that the entire webroot and all it's
contents had the correct user and group.

From the rsync command, I removed the permissions ignoring and timestamp
ignoring, so I can keep the permissions that are now correctly set on
the server, regardless of the source.

The timestamp preservation allows for incremental syncs; before, the
command would delete everything on the server, then sync everything on
the source. Now, only what's changed will be synced.

I excluded dotfiles from the sync, so we don't get weird and possibly
insecure files open to the public.

I suppressed output from the command to only show errors, so I'm not
distracted by unneccessary output, and my Neovim Make command doesn't
always fill up the quickfix list with noise.

I used the long name for options for readability.