Simplify article display
Dispense with the "featured" vs "summary" article distinctions on the
home page. Only print the first three, have them all include the
summary, and change them to use headlines. That looks nice and is
semantic.

The article listing page can still be bare with the link list and all
that.