Update micro feed output
Output the content title and type for micro feeds. This indicates that
this isn't a regular note post, and also hints at the type of content
linked.

Use the Relative permalink for the GUID for both the regular feed and
the micro feed. Copying marcus.io. Should still be unique, but without
extra bytes that will always be identical.