Fixup code block font sizes
The font sizing for code blocks is inconsistent across tags. This is
because they're sized in bulk via `em`s, which compound because the tags
in question - `<pre>`, `<code>`, `<kbd>` - are frequently nested within
each other.

Style each tag separately when appropriate, and use `rem` here to keep
everything consistent across nesting. Also add a little rounded edge to
blocks for a flourish.