Defeat general margin for svg links + nested lists
There is a general whitespace rule applied to all elements:

    * + * { margin-top: 1em }

This usually gives a hands-free, pleasing vertical rhythm, but in some
cases requires overrides.

With icons nested inside links, we don't want that extra margin applied
to the svg, since that leads to inconsistent line heights.

We also don't want nested `ul`s getting the extra whitespace, since
those should be grouped to it's parent, not spaced out as if it were
separate.