Add aside shortcode
I use the `aside` tag several times to call out related information. I'm
using it a third time in an upcoming article, so codify it as a pattern
in a shortcode. Ditch the arbitrary markup support inside headings, as
it's only being used by the `abbr` tag, which in context is pretty
unhelpful anyway.