Remove fancy category styles
These assume there will only ever be four categories. This has changed.
Instead of trying to keep up with a brittle style, or refactor to some
other "fancy" style that accommodates more than four categories, keep it
as a simple list. This page is not prominently placed right now anyway.